renaming of the files for better usage
added the collection_explanation_methods.py
