+++ python/collection_loss_functions.py
